Refactor: extract About view into dedicated module
Move About modal rendering from modal_text.py to ui/about_view.py.

Reduce responsibility of ModalTextBuilder and isolate About-specific
UI logic in its own module.

No behavior change.
